test: split machine-k8s agents tests
