+++ tests/integration/test_k8s_agent.py
@@ -7,6 +7,7 @@ import logging
 from pathlib import Path
 
 import jenkinsapi.jenkins
+import pytest
 import requests
 import yaml
 from juju.application import Application
@@ -67,41 +68,13 @@ async def test_jenkins_k8s_agent_relation(
     assert not any((application.name in key for key in jenkins_client.get_nodes().keys()))
 
 
-async def test_jenkins_machine_agent_relation(
-    application: Application,
-    jenkins_machine_agents: Application,
-    jenkins_client: jenkinsapi.jenkins.Jenkins,
+@pytest.mark.usefixtures("app_k8s_deprecated_agent_related")
+async def test_jenkins_k8s_deprecated_agent_relation(
+    jenkins_k8s_agents: Application, jenkins_client: jenkinsapi.jenkins.Jenkins
 ):
     """
-    arrange: given a cross controller cross model jenkins machine agent with an offer.
-    act:
-        1. when the server charm is related to the machine agent charm.
-        2. when the relation is removed.
-    assert:
-        1. the relation succeeds and the machine agent is able to run jobs successfully.
-        2. the machine agent is deregistered from Jenkins.
+    arrange: given jenkins-k8s-agent and jenkins server charms.
+    act: when the server charm is related to the k8s agent charm.
+    assert: the relation succeeds and the agent is able to run jobs successfully.
     """
-    # 1. Relate jenkins-k8s charm to the jenkins-agent charm.
-    model = application.model
-    machine_model = jenkins_machine_agents.model
-    # this code is similar to the app_machine_agent_related fixture but shouldn't be using the
-    # fixture since this test tests for teardown of relation as well.
-    # pylint: disable=duplicate-code
-    await model.relate(
-        f"{application.name}:{state.AGENT_RELATION}",
-        f"localhost:admin/{machine_model.name}.{state.AGENT_RELATION}",
-    )
-    await machine_model.wait_for_idle(apps=[jenkins_machine_agents.name], wait_for_active=True)
-    await model.wait_for_idle(apps=[application.name], wait_for_active=True)
-    # pylint: enable=duplicate-code
-
-    # 1. Assert that the node is registered and is able to run jobs successfully.
-    assert_job_success(jenkins_client, jenkins_machine_agents.name, "machine")
-
-    # 2. Remove the relation
-    await application.remove_relation(state.AGENT_RELATION, state.AGENT_RELATION)
-    await model.wait_for_idle(apps=[application.name])
-    await machine_model.wait_for_idle(apps=[jenkins_machine_agents.name])
-
-    # 2. Assert that the agent nodes are deregistered from Jenkins.
-    assert not any((application.name in key for key in jenkins_client.get_nodes().keys()))
+    assert_job_success(jenkins_client, jenkins_k8s_agents.name, "k8s")
